backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
openai==1.12.0
anthropic==0.18.1
google-auth==2.27.0
google-api-python-client==2.115.0
sendgrid==6.11.0
pytest==7.4.4
pytest-asyncio==0.23.4
